file-scope: Extract single-file batch discard
The multi-file discard-to-batch support fell back through the discard
command entry point for files outside the precollected text path.

That dependency pointed support code back at command dispatch and kept
whole-file binary and deletion handling inside the entry module.

This commit moves the single-file batch discard fallback into file-scope
support and routes both callers through that module.

A following tests commit will pin the new boundary.
