+++ src/git_stage_batch/commands/discard.py
@@ -10,13 +10,9 @@ from contextlib import ExitStack
 
 from ..batch.operations import create_batch
 from ..batch.storage import (
-    add_binary_file_to_batch,
     add_file_to_batch,
 )
 from ..batch.display import annotate_with_batch_source
-from ..batch.ownership import (
-    BatchOwnership,
-)
 from ..core.replacement import (
     ReplacementPayload,
     coerce_replacement_payload,
@@ -38,8 +34,6 @@ from ..core.hashing import (
     compute_text_file_deletion_hash,
 )
 from ..core.models import BinaryFileChange, GitlinkChange, RenameChange, TextFileDeletionChange
-from ..core.text_lifecycle import TextFileChangeType
-from ..data.text_lifecycle_detection import detect_empty_text_lifecycle_change
 from ..data.hunk_tracking import (
     fetch_next_change,
 )
@@ -59,7 +53,6 @@ from ..data.selected_change.clear_reasons import (
     refuse_bare_action_after_file_list,
 )
 from ..data.file_change_display import (
-    render_binary_file_change,
     render_gitlink_change,
     render_rename_change,
     render_text_deletion_change,
@@ -119,6 +112,11 @@ from .selection import discard_line_selection as _discard_line_selection
 from .selection import discard_line_replacement as _discard_line_replacement
 from .selection import batch_line_selection as _batch_line_selection
 from .selection import batch_line_updates as _batch_line_updates
+from .file_scope.discard_file_to_batch import (
+    discard_binary_to_batch,
+    discard_file_to_batch,
+    discard_text_deletion_to_batch,
+)
 from .selection.selected_hunk_refresh import (
     recalculate_selected_hunk_for_command,
     refresh_selected_hunk_after_line_action,
@@ -712,7 +710,7 @@ def command_discard_to_batch(
         ):
             selected_change = load_selected_change()
             if isinstance(selected_change, BinaryFileChange):
-                saved_hunks = _command_discard_binary_to_batch(
+                saved_hunks = discard_binary_to_batch(
                     batch_name,
                     selected_change,
                     quiet=quiet,
@@ -732,7 +730,7 @@ def command_discard_to_batch(
         ):
             selected_change = load_selected_change()
             if isinstance(selected_change, TextFileDeletionChange):
-                saved_hunks = _command_discard_text_deletion_to_batch(
+                saved_hunks = discard_text_deletion_to_batch(
                     batch_name,
                     selected_change,
                     quiet=quiet,
@@ -759,7 +757,7 @@ def command_discard_to_batch(
 
             if line_ids is None:
                 # --file without --line: discard entire file
-                saved_hunks = _command_discard_file_to_batch(
+                saved_hunks = discard_file_to_batch(
                     batch_name,
                     target_file,
                     quiet=quiet,
@@ -924,297 +922,6 @@ def _command_discard_lines_to_batch_as(
         replacement.file_path,
         auto_advance=auto_advance,
     )
-
-
-def _discard_binary_change_from_working_tree(binary_change: BinaryFileChange) -> None:
-    """Discard one live binary change from the working tree."""
-    file_path = binary_change.new_path if binary_change.new_path != "/dev/null" else binary_change.old_path
-    absolute_path = get_git_repository_root_path() / file_path
-
-    if binary_change.is_new_file():
-        if absolute_path.exists():
-            absolute_path.unlink()
-        git_remove_paths([file_path], cached=True, quiet=True, check=False)
-        return
-
-    result = git_checkout_paths("HEAD", [file_path], check=False)
-    if result.returncode != 0:
-        exit_with_error(_("Failed to restore binary file: {error}").format(error=result.stderr))
-
-
-def _command_discard_binary_to_batch(
-    batch_name: str,
-    binary_change: BinaryFileChange,
-    *,
-    quiet: bool = False,
-    advance: bool = True,
-    auto_advance: bool | None = None,
-) -> int:
-    """Save one binary change to a batch, then discard it from the working tree."""
-    file_path = binary_change.new_path if binary_change.new_path != "/dev/null" else binary_change.old_path
-    patch_hash = compute_binary_file_hash(binary_change)
-
-    snapshot_file_if_untracked(file_path)
-    add_binary_file_to_batch(
-        batch_name,
-        binary_change,
-        file_mode=detect_file_mode(file_path),
-    )
-    _discard_binary_change_from_working_tree(binary_change)
-
-    append_lines_to_file(get_block_list_file_path(), [patch_hash])
-    record_hunk_discarded(patch_hash)
-
-    if not quiet:
-        print(
-            _("Discarded binary file '{file}' to batch '{batch}'").format(
-                file=file_path,
-                batch=batch_name,
-            ),
-            file=sys.stderr,
-        )
-
-    if advance:
-        finish_selected_change_action(quiet=quiet, auto_advance=auto_advance)
-    return 1
-
-
-def _command_discard_text_deletion_to_batch(
-    batch_name: str,
-    deletion_change: TextFileDeletionChange,
-    *,
-    quiet: bool = False,
-    advance: bool = True,
-    auto_advance: bool | None = None,
-) -> int:
-    """Save one whole-text-file deletion to a batch, then restore it locally."""
-    file_path = deletion_change.path()
-    patch_hash = compute_text_file_deletion_hash(deletion_change)
-
-    snapshot_file_if_untracked(file_path)
-    add_file_to_batch(
-        batch_name,
-        file_path,
-        BatchOwnership([], []),
-        detect_file_mode(file_path),
-        change_type=TextFileChangeType.DELETED.value,
-    )
-    discard_text_deletion_change(deletion_change)
-
-    append_lines_to_file(get_block_list_file_path(), [patch_hash])
-    record_hunk_discarded(patch_hash)
-
-    if not quiet:
-        print(
-            _("Discarded text file deletion '{file}' to batch '{batch}'").format(
-                file=file_path,
-                batch=batch_name,
-            ),
-            file=sys.stderr,
-        )
-
-    if advance:
-        finish_selected_change_action(quiet=quiet, auto_advance=auto_advance)
-    return 1
-
-
-def _command_discard_file_to_batch(
-    batch_name: str,
-    file_path: str,
-    *,
-    quiet: bool = False,
-    advance: bool = True,
-    auto_advance: bool | None = None,
-) -> int:
-    """Discard entire file to batch (internal helper for file-scoped operations)."""
-    auto_add_untracked_files([file_path])
-
-    log_journal("discard_file_to_batch_start", batch_name=batch_name, file_path=file_path, quiet=quiet)
-
-    # Auto-create batch if it doesn't exist
-    if not batch_exists(batch_name):
-        create_batch(batch_name, "Auto-created")
-
-    deletion_change = render_text_deletion_change(file_path)
-    if deletion_change is not None:
-        return _command_discard_text_deletion_to_batch(
-            batch_name,
-            deletion_change,
-            quiet=quiet,
-            advance=advance,
-            auto_advance=auto_advance,
-        )
-
-    binary_change = render_binary_file_change(file_path)
-    if binary_change is not None:
-        return _command_discard_binary_to_batch(
-            batch_name,
-            binary_change,
-            quiet=quiet,
-            advance=advance,
-            auto_advance=auto_advance,
-        )
-    if render_gitlink_change(file_path) is not None:
-        exit_with_error(_("Discarding submodule pointer changes to a batch is not supported yet."))
-
-    # Load blocklist
-    blocklist_path = get_block_list_file_path()
-    blocked_hashes = read_text_file_line_set(blocklist_path)
-
-    # Detect file mode
-    file_mode = detect_file_mode(file_path)
-
-    with ExitStack() as patch_stack:
-        # Collect ALL hunks from this file (live working tree state)
-        all_lines_to_batch = []
-        patches_to_discard = []
-
-        with acquire_unified_diff(
-            stream_live_git_diff(
-                base="HEAD",
-                context_lines=get_context_lines(),
-                paths=[file_path],
-            )
-        ) as patches:
-            for patch in patches:
-                if isinstance(patch, RenameChange):
-                    continue
-
-                if isinstance(patch, TextFileDeletionChange):
-                    continue
-
-                if isinstance(patch, GitlinkChange):
-                    exit_with_error(_("Discarding submodule pointer changes to a batch is not supported yet."))
-                if isinstance(patch, BinaryFileChange):
-                    continue
-
-                patch_hash = compute_stable_hunk_hash_from_lines(patch.lines)
-
-                if patch_hash in blocked_hashes:
-                    continue
-
-                # Parse hunk to get lines
-                hunk_lines = build_line_changes_from_patch_lines(
-                    patch.lines,
-                    annotator=annotate_with_batch_source,
-                )
-                all_lines_to_batch.extend(hunk_lines.lines)
-                patches_to_discard.append((
-                    patch_stack.enter_context(LineBuffer.from_chunks(patch.lines)),
-                    patch_hash,
-                ))
-
-        if not all_lines_to_batch:
-            # Special case: empty text lifecycle changes have no hunk body.
-            repo_root = get_git_repository_root_path()
-            full_path = repo_root / file_path
-            lifecycle_change_type = detect_empty_text_lifecycle_change(file_path)
-            if lifecycle_change_type is not None:
-                snapshot_file_if_untracked(file_path)
-                add_file_to_batch(
-                    batch_name,
-                    file_path,
-                    BatchOwnership([], []),
-                    file_mode,
-                    change_type=lifecycle_change_type,
-                )
-
-                if lifecycle_change_type == TextFileChangeType.ADDED:
-                    # Delete from working tree
-                    full_path.unlink()
-
-                    # Remove from index if present (from intent-to-add)
-                    git_remove_paths([file_path], cached=True, quiet=True, check=False)
-                else:
-                    git_checkout_paths("HEAD", [file_path], check=False)
-
-                if not quiet:
-                    print(_("Discarded file '{file}' to batch '{batch}'").format(file=file_path, batch=batch_name), file=sys.stderr)
-
-                log_journal("discard_file_to_batch_end", batch_name=batch_name, file_path=file_path)
-                return 1
-
-            if not quiet:
-                print(_("No changes in file '{file}' to discard.").format(file=file_path), file=sys.stderr)
-            return 0
-
-        # Prepare batch ownership update (handles stale source, translation, merge)
-
-        metadata = read_batch_metadata(batch_name)
-        file_metadata = metadata.get("files", {}).get(file_path)
-
-        with ExitStack() as ownership_stack:
-            try:
-                update = ownership_stack.enter_context(
-                    acquire_batch_ownership_update_for_selection(
-                        batch_name=batch_name,
-                        file_path=file_path,
-                        file_metadata=file_metadata,
-                        selected_lines=all_lines_to_batch,
-                    )
-                )
-            except ValueError as e:
-                exit_with_error(
-                    _("Cannot discard file to batch: batch source is stale and remapping failed.\n"
-                      "File: {file}\n"
-                      "Batch: {batch}\n"
-                      "Error: {error}").format(file=file_path, batch=batch_name, error=str(e))
-                )
-
-            # Snapshot file before modifying
-            snapshot_file_if_untracked(file_path)
-
-            # Save to batch
-            add_file_to_batch(
-                batch_name,
-                file_path,
-                update.ownership_after,
-                file_mode,
-                batch_source_commit=update.batch_source_commit,
-            )
-
-        # Record hunks as discarded for progress tracking
-        for _patch_lines, patch_hash in patches_to_discard:
-            record_hunk_discarded(patch_hash)
-
-        # Discard from working tree (reverse patches)
-        for patch_lines_item, patch_hash in patches_to_discard:
-            log_journal(
-                "discard_file_to_batch_before_git_apply",
-                batch_name=batch_name,
-                patch_hash=patch_hash,
-                file_path=file_path,
-                patch_line_count=len(patch_lines_item),
-            )
-
-            apply_result = git_apply_to_worktree(
-                patch_lines_item,
-                reverse=True,
-                unidiff_zero=True,
-                check=False,
-            )
-
-            if apply_result.returncode != 0:
-                exit_with_error(_("Failed to discard changes from file: {err}").format(err=apply_result.stderr))
-
-            log_journal("discard_file_to_batch_after_git_apply", batch_name=batch_name, patch_hash=patch_hash, exit_code=apply_result.returncode)
-
-        # If file was deleted from working tree, remove from index too
-        repo_root = get_git_repository_root_path()
-        full_path = repo_root / file_path
-        if not full_path.exists():
-            # File was deleted by git apply --reverse, remove from index
-            git_remove_paths([file_path], cached=True, quiet=True, check=False)
-
-        if not quiet:
-            print(_("Discarded file '{file}' to batch '{batch}'").format(file=file_path, batch=batch_name), file=sys.stderr)
-
-        log_journal("discard_file_to_batch_end", batch_name=batch_name, file_path=file_path)
-
-        # Show next hunk
-        if advance:
-            finish_selected_change_action(quiet=quiet, auto_advance=auto_advance)
-        return len(patches_to_discard)
 
 
 def _command_discard_file_lines_to_batch(
@@ -1391,14 +1098,14 @@ def _command_discard_hunk_to_batch(
     if isinstance(selected_item, GitlinkChange):
         exit_with_error(_("Discarding submodule pointer changes to a batch is not supported yet."))
     if isinstance(selected_item, BinaryFileChange):
-        return _command_discard_binary_to_batch(
+        return discard_binary_to_batch(
             batch_name,
             selected_item,
             quiet=quiet,
             auto_advance=auto_advance,
         )
     if isinstance(selected_item, TextFileDeletionChange):
-        return _command_discard_text_deletion_to_batch(
+        return discard_text_deletion_to_batch(
             batch_name,
             selected_item,
             quiet=quiet,
