+++ src/git_stage_batch/commands/file_scope/discard_file_to_batch.py
@@ -1,0 +1,336 @@
+"""Single-file discard-to-batch support."""
+
+from __future__ import annotations
+
+from contextlib import ExitStack
+import sys
+
+from ...batch.display import annotate_with_batch_source
+from ...batch.operations import create_batch
+from ...batch.ownership import BatchOwnership
+from ...batch.query import read_batch_metadata
+from ...batch.source_refresh import acquire_batch_ownership_update_for_selection
+from ...batch.storage import add_binary_file_to_batch, add_file_to_batch
+from ...batch.validation import batch_exists
+from ...core.buffer import LineBuffer
+from ...core.diff_parser import acquire_unified_diff, build_line_changes_from_patch_lines
+from ...core.hashing import (
+    compute_binary_file_hash,
+    compute_stable_hunk_hash_from_lines,
+    compute_text_file_deletion_hash,
+)
+from ...core.models import BinaryFileChange, GitlinkChange, RenameChange, TextFileDeletionChange
+from ...core.text_lifecycle import TextFileChangeType
+from ...data.file_change_display import (
+    render_binary_file_change,
+    render_gitlink_change,
+    render_text_deletion_change,
+)
+from ...data.file_modes import detect_file_mode
+from ...data.file_tracking import auto_add_untracked_files
+from ...data.live_diff import stream_live_git_diff
+from ...data.progress import record_hunk_discarded
+from ...data.session import snapshot_file_if_untracked
+from ...data.text_lifecycle_detection import detect_empty_text_lifecycle_change
+from ...exceptions import exit_with_error
+from ...i18n import _
+from ...utils.file_io import append_lines_to_file, read_text_file_line_set
+from ...utils.git import (
+    get_git_repository_root_path,
+    git_apply_to_worktree,
+    git_checkout_paths,
+    git_remove_paths,
+)
+from ...utils.journal import log_journal
+from ...utils.paths import get_block_list_file_path, get_context_lines
+from ..selection.action_completion import finish_selected_change_action
+from ..selection.selected_change_discarding import discard_text_deletion_change
+
+
+def _discard_binary_change_from_working_tree(binary_change: BinaryFileChange) -> None:
+    """Discard one live binary change from the working tree."""
+    file_path = binary_change.new_path if binary_change.new_path != "/dev/null" else binary_change.old_path
+    absolute_path = get_git_repository_root_path() / file_path
+
+    if binary_change.is_new_file():
+        if absolute_path.exists():
+            absolute_path.unlink()
+        git_remove_paths([file_path], cached=True, quiet=True, check=False)
+        return
+
+    result = git_checkout_paths("HEAD", [file_path], check=False)
+    if result.returncode != 0:
+        exit_with_error(_("Failed to restore binary file: {error}").format(error=result.stderr))
+
+
+def discard_binary_to_batch(
+    batch_name: str,
+    binary_change: BinaryFileChange,
+    *,
+    quiet: bool = False,
+    advance: bool = True,
+    auto_advance: bool | None = None,
+) -> int:
+    """Save one binary change to a batch, then discard it from the working tree."""
+    file_path = binary_change.new_path if binary_change.new_path != "/dev/null" else binary_change.old_path
+    patch_hash = compute_binary_file_hash(binary_change)
+
+    snapshot_file_if_untracked(file_path)
+    add_binary_file_to_batch(
+        batch_name,
+        binary_change,
+        file_mode=detect_file_mode(file_path),
+    )
+    _discard_binary_change_from_working_tree(binary_change)
+
+    append_lines_to_file(get_block_list_file_path(), [patch_hash])
+    record_hunk_discarded(patch_hash)
+
+    if not quiet:
+        print(
+            _("Discarded binary file '{file}' to batch '{batch}'").format(
+                file=file_path,
+                batch=batch_name,
+            ),
+            file=sys.stderr,
+        )
+
+    if advance:
+        finish_selected_change_action(quiet=quiet, auto_advance=auto_advance)
+    return 1
+
+
+def discard_text_deletion_to_batch(
+    batch_name: str,
+    deletion_change: TextFileDeletionChange,
+    *,
+    quiet: bool = False,
+    advance: bool = True,
+    auto_advance: bool | None = None,
+) -> int:
+    """Save one whole-text-file deletion to a batch, then restore it locally."""
+    file_path = deletion_change.path()
+    patch_hash = compute_text_file_deletion_hash(deletion_change)
+
+    snapshot_file_if_untracked(file_path)
+    add_file_to_batch(
+        batch_name,
+        file_path,
+        BatchOwnership([], []),
+        detect_file_mode(file_path),
+        change_type=TextFileChangeType.DELETED.value,
+    )
+    discard_text_deletion_change(deletion_change)
+
+    append_lines_to_file(get_block_list_file_path(), [patch_hash])
+    record_hunk_discarded(patch_hash)
+
+    if not quiet:
+        print(
+            _("Discarded text file deletion '{file}' to batch '{batch}'").format(
+                file=file_path,
+                batch=batch_name,
+            ),
+            file=sys.stderr,
+        )
+
+    if advance:
+        finish_selected_change_action(quiet=quiet, auto_advance=auto_advance)
+    return 1
+
+
+def discard_file_to_batch(
+    batch_name: str,
+    file_path: str,
+    *,
+    quiet: bool = False,
+    advance: bool = True,
+    auto_advance: bool | None = None,
+) -> int:
+    """Discard one file to a batch."""
+    auto_add_untracked_files([file_path])
+
+    log_journal("discard_file_to_batch_start", batch_name=batch_name, file_path=file_path, quiet=quiet)
+
+    if not batch_exists(batch_name):
+        create_batch(batch_name, "Auto-created")
+
+    deletion_change = render_text_deletion_change(file_path)
+    if deletion_change is not None:
+        return discard_text_deletion_to_batch(
+            batch_name,
+            deletion_change,
+            quiet=quiet,
+            advance=advance,
+            auto_advance=auto_advance,
+        )
+
+    binary_change = render_binary_file_change(file_path)
+    if binary_change is not None:
+        return discard_binary_to_batch(
+            batch_name,
+            binary_change,
+            quiet=quiet,
+            advance=advance,
+            auto_advance=auto_advance,
+        )
+    if render_gitlink_change(file_path) is not None:
+        exit_with_error(_("Discarding submodule pointer changes to a batch is not supported yet."))
+
+    blocklist_path = get_block_list_file_path()
+    blocked_hashes = read_text_file_line_set(blocklist_path)
+    file_mode = detect_file_mode(file_path)
+
+    with ExitStack() as patch_stack:
+        all_lines_to_batch = []
+        patches_to_discard = []
+
+        with acquire_unified_diff(
+            stream_live_git_diff(
+                base="HEAD",
+                context_lines=get_context_lines(),
+                paths=[file_path],
+            )
+        ) as patches:
+            for patch in patches:
+                if isinstance(patch, RenameChange):
+                    continue
+
+                if isinstance(patch, TextFileDeletionChange):
+                    continue
+
+                if isinstance(patch, GitlinkChange):
+                    exit_with_error(_("Discarding submodule pointer changes to a batch is not supported yet."))
+                if isinstance(patch, BinaryFileChange):
+                    continue
+
+                patch_hash = compute_stable_hunk_hash_from_lines(patch.lines)
+
+                if patch_hash in blocked_hashes:
+                    continue
+
+                hunk_lines = build_line_changes_from_patch_lines(
+                    patch.lines,
+                    annotator=annotate_with_batch_source,
+                )
+                all_lines_to_batch.extend(hunk_lines.lines)
+                patches_to_discard.append((
+                    patch_stack.enter_context(LineBuffer.from_chunks(patch.lines)),
+                    patch_hash,
+                ))
+
+        if not all_lines_to_batch:
+            repo_root = get_git_repository_root_path()
+            full_path = repo_root / file_path
+            lifecycle_change_type = detect_empty_text_lifecycle_change(file_path)
+            if lifecycle_change_type is not None:
+                snapshot_file_if_untracked(file_path)
+                add_file_to_batch(
+                    batch_name,
+                    file_path,
+                    BatchOwnership([], []),
+                    file_mode,
+                    change_type=lifecycle_change_type,
+                )
+
+                if lifecycle_change_type == TextFileChangeType.ADDED:
+                    full_path.unlink()
+                    git_remove_paths([file_path], cached=True, quiet=True, check=False)
+                else:
+                    git_checkout_paths("HEAD", [file_path], check=False)
+
+                if not quiet:
+                    print(
+                        _("Discarded file '{file}' to batch '{batch}'").format(
+                            file=file_path,
+                            batch=batch_name,
+                        ),
+                        file=sys.stderr,
+                    )
+
+                log_journal("discard_file_to_batch_end", batch_name=batch_name, file_path=file_path)
+                return 1
+
+            if not quiet:
+                print(_("No changes in file '{file}' to discard.").format(file=file_path), file=sys.stderr)
+            return 0
+
+        metadata = read_batch_metadata(batch_name)
+        file_metadata = metadata.get("files", {}).get(file_path)
+
+        with ExitStack() as ownership_stack:
+            try:
+                update = ownership_stack.enter_context(
+                    acquire_batch_ownership_update_for_selection(
+                        batch_name=batch_name,
+                        file_path=file_path,
+                        file_metadata=file_metadata,
+                        selected_lines=all_lines_to_batch,
+                    )
+                )
+            except ValueError as e:
+                exit_with_error(
+                    _("Cannot discard file to batch: batch source is stale and remapping failed.\n"
+                      "File: {file}\n"
+                      "Batch: {batch}\n"
+                      "Error: {error}").format(file=file_path, batch=batch_name, error=str(e))
+                )
+
+            snapshot_file_if_untracked(file_path)
+
+            add_file_to_batch(
+                batch_name,
+                file_path,
+                update.ownership_after,
+                file_mode,
+                batch_source_commit=update.batch_source_commit,
+            )
+
+        for _patch_lines, patch_hash in patches_to_discard:
+            record_hunk_discarded(patch_hash)
+
+        for patch_lines_item, patch_hash in patches_to_discard:
+            log_journal(
+                "discard_file_to_batch_before_git_apply",
+                batch_name=batch_name,
+                patch_hash=patch_hash,
+                file_path=file_path,
+                patch_line_count=len(patch_lines_item),
+            )
+
+            apply_result = git_apply_to_worktree(
+                patch_lines_item,
+                reverse=True,
+                unidiff_zero=True,
+                check=False,
+            )
+
+            if apply_result.returncode != 0:
+                exit_with_error(_("Failed to discard changes from file: {err}").format(err=apply_result.stderr))
+
+            log_journal(
+                "discard_file_to_batch_after_git_apply",
+                batch_name=batch_name,
+                patch_hash=patch_hash,
+                exit_code=apply_result.returncode,
+            )
+
+        repo_root = get_git_repository_root_path()
+        full_path = repo_root / file_path
+        if not full_path.exists():
+            git_remove_paths([file_path], cached=True, quiet=True, check=False)
+
+        if not quiet:
+            print(
+                _("Discarded file '{file}' to batch '{batch}'").format(
+                    file=file_path,
+                    batch=batch_name,
+                ),
+                file=sys.stderr,
+            )
+
+        log_journal("discard_file_to_batch_end", batch_name=batch_name, file_path=file_path)
+
+        if advance:
+            finish_selected_change_action(quiet=quiet, auto_advance=auto_advance)
+        return len(patches_to_discard)
